Docs, got rid of some globals
