--- src/consumers/shared/shared.c
@@ -1,7 +1,0 @@
-#include "../../statsd-parsers/statsd-parsers.h"
-#include "./shared.h"
-
-metric_metadata* create_record_meta(statsd_datagram* datagram) {
-    // TODO: this is not yet implemented
-    return NULL;
-}
